+++ crates/query-engine/translation/src/translation/mutation/v2/generate.rs
@@ -16,21 +16,27 @@ pub enum Mutation {
 }
 
 /// Given our introspection data, work out all the mutations we can generate
-pub fn generate(tables_info: &database::TablesInfo) -> BTreeMap<models::ProcedureName, Mutation> {
+pub fn generate(
+    tables_info: &database::TablesInfo,
+    mutations_prefix: &Option<String>,
+) -> BTreeMap<models::ProcedureName, Mutation> {
     let mut mutations = BTreeMap::new();
     for (collection_name, table_info) in &tables_info.0 {
         // Delete mutations.
-        let delete_mutations = generate_delete_by_unique(collection_name, table_info);
+        let delete_mutations =
+            generate_delete_by_unique(collection_name, table_info, mutations_prefix);
         for (name, delete_mutation) in delete_mutations {
             mutations.insert(name, Mutation::DeleteMutation(delete_mutation));
         }
 
         // Insert mutations.
-        let (name, insert_mutation) = insert::generate(collection_name, table_info);
+        let (name, insert_mutation) =
+            insert::generate(collection_name, table_info, mutations_prefix);
         mutations.insert(name, Mutation::InsertMutation(insert_mutation));
 
         // Update mutations.
-        let update_mutations = generate_update_by_unique(collection_name, table_info);
+        let update_mutations =
+            generate_update_by_unique(collection_name, table_info, mutations_prefix);
         for (name, update_mutation) in update_mutations {
             mutations.insert(name, Mutation::UpdateMutation(update_mutation));
         }