+++ changelog.md
@@ -8,6 +8,9 @@
 
 - Change mutation pre/post checks to be optional
 
+- Added `mutationPrefix` to configuration allowing mutation names to be
+  customised.
+
 ### Fixed
 
 ## [v1.1.2] - 2024-10-07